Backend/requirements.txt
asgiref==3.8.1
bcc==0.1.10
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
numpy==2.2.4
PyMySQL==1.1.1
sqlparse==0.5.3
traitlets==5.14.3
traittypes==0.2.1
